implement plugins for auth
