+++ enterprise/frontend/src/embedding-sdk-bundle/test/auth-flow/jwt.unit.spec.tsx
@@ -23,6 +23,9 @@ import {
 
 import { setup as baseSetup } from "./setup";
 
+// Activates the plugins needed for the tests before any SDK code
+import "metabase-enterprise/sdk-plugins";
+
 const setup = ({
   authConfig,
   locale,
